I've been neglecting committing, but a lot of progress on the qt transition here. using qt thread for exp thread, but still seems like some minor issues with thread crashing with variables?

+++ Chimera/Source/GeneralObjects/ExpWrap.h
@@ -28,7 +28,7 @@ class ExpWrap
 			return data[ variationNumber ];
 		};
 
-		UINT getNumVariations(  )
+		size_t getNumVariations(  )
 		{
 			return data.size();
 		}
